# openapi.yaml
openapi: 3.1.0
info:
    title: Terraform HTTP State Backend
    version: '1.0.0'
    summary: >
        Wire protocol implemented by a server that acts as a state backend for the
        Terraform (and OpenTofu) `http` backend.
    description: |
        This spec documents the HTTP contract the Terraform CLI expects when
        configured with `backend "http"`. It is reconstructed from the Terraform
        source (see `x-sources` below) because the official documentation page
        lists configuration options but not request/response schemas.

        ## Paths in this spec

        A real deployment typically exposes two URLs:
          - A **state URL** (`address`) used for `GET`, the update method
            (`POST` by default, often `PUT`), and `DELETE`.
          - A **lock URL** (`lock_address` / `unlock_address`) used for
            `LOCK`/`UNLOCK` — in practice these are almost always overridden to
            `POST` and `DELETE` on a `.../lock` sub-path.

        Both URLs may be the same resource (and are on some servers such as
        GitLab's), in which case the HTTP method disambiguates the operation.

        ## Authentication

        The CLI sends credentials as HTTP Basic Auth when `username` and
        `password` (or `TF_HTTP_USERNAME` / `TF_HTTP_PASSWORD`) are set. Mutual
        TLS is also supported via `client_certificate_pem` etc. There is no
        provision in the protocol itself for Bearer tokens — servers that need
        bearer auth typically accept the token as the Basic password.

    x-sources:
        - description: Backend config schema and env-var mapping
          url: https://github.com/hashicorp/terraform/blob/main/internal/backend/remote-state/http/backend.go
        - description: HTTP client — authoritative implementation of the wire protocol (Get, Put, Delete, Lock, Unlock)
          url: https://github.com/hashicorp/terraform/blob/main/internal/backend/remote-state/http/client.go
        - description: LockInfo struct definition (JSON field names and types)
          url: https://github.com/hashicorp/terraform/blob/main/internal/states/statemgr/locker.go
        - description: Original PR that introduced locking support (#15793)
          url: https://github.com/hashicorp/terraform/pull/15793
        - description: Official backend documentation (config options only)
          url: https://developer.hashicorp.com/terraform/language/backend/http
        - description: Real-world server implementation — GitLab managed Terraform state
          url: https://gitlab.com/gitlab-org/gitlab/-/tree/master/app/controllers/api/v4/terraform
        - description: Known issue — force-unlock sends an empty body (#28421)
          url: https://github.com/hashicorp/terraform/issues/28421
        - description: Known issue — lock_method=PUT expects 200/201, not 204 (#32731)
          url: https://github.com/hashicorp/terraform/issues/32731

servers:
    - url: https://state.example.com
      description: Your state server

paths:
    /state/{name}:
        parameters:
            - $ref: '#/components/parameters/StateName'

        get:
            summary: Fetch current state
            description: |
                Returns the raw Terraform state JSON. Called on `terraform init`,
                `plan`, `apply`, `refresh`, and any read.

                **No body or a 404/204 both mean "no state yet"** — Terraform will
                start from an empty state. Returning `200` with an empty body is
                **not** valid JSON and will cause a parse error in the CLI.
            responses:
                '200':
                    description: Current state snapshot.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TerraformState'
                '204':
                    description: No state stored yet. Treated as empty state by the CLI.
                '404':
                    description: No state stored yet. Treated as empty state by the CLI.
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'

        post:
            summary: Update state (default update_method)
            description: |
                Persists a new state snapshot. The full state JSON is sent — this is
                a replace, not a diff. The CLI sends `Content-MD5` (base64 of the
                raw MD5 digest of the body); the server may verify it but is not
                required to.

                If locking is enabled, the request carries the `ID` query parameter
                containing the lock ID that was returned from a prior `LOCK`. The
                server **must** reject the write if this ID does not match the
                currently held lock.
            parameters:
                - $ref: '#/components/parameters/LockID'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TerraformState'
            responses:
                '200':
                    description: State persisted.
                '204':
                    description: State persisted (no body).
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '409':
                    description: Lock ID mismatch — the supplied `ID` does not own the current lock.
                '5XX':
                    description: >
                        Server error. The CLI retries 5xx responses up to `retry_max`
                        times (default 2), with exponential backoff between
                        `retry_wait_min` and `retry_wait_max`.

        put:
            summary: Update state (when update_method = "PUT")
            description: |
                Same semantics as `POST`. Use this only when your server was
                configured with `update_method = "PUT"` on the client side.
            parameters:
                - $ref: '#/components/parameters/LockID'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TerraformState'
            responses:
                '200':
                    description: State persisted.

        delete:
            summary: Delete state
            description: |
                Called by `terraform workspace delete` and similar. Optional — if you
                don't implement it, return `405 Method Not Allowed`; `apply`/`plan`
                flows will still work.
            responses:
                '200':
                    description: State deleted.
                '204':
                    description: State deleted (no body).
                '404':
                    description: Nothing to delete. Treat as success.
                '405':
                    description: Not implemented.

    /state/{name}/lock:
        parameters:
            - $ref: '#/components/parameters/StateName'

        post:
            summary: Acquire lock (when lock_method = "POST")
            description: |
                Acquires an exclusive lock on the given state. The request body is
                the `LockInfo` struct the CLI generated for this operation.

                **The 409/423 response body is part of the protocol** — it must
                contain the currently held `LockInfo` as JSON. The CLI parses it
                and renders the "Lock Info:" block shown to the user. Returning a
                plain-text error here causes the confusing
                `failed to unmarshal body` message.

                The default HTTP verb is `LOCK`, but virtually every real deployment
                overrides `lock_method` to `POST` because `LOCK` is a WebDAV verb
                and awkward to route. Document both in your API.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LockInfo'
            responses:
                '200':
                    description: Lock acquired.
                '409':
                    description: Already locked. Body is the current holder's LockInfo.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LockInfo'
                '423':
                    description: Already locked (alternative status). Body is the current holder's LockInfo.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LockInfo'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'

        delete:
            summary: Release lock (when unlock_method = "DELETE")
            description: |
                Releases the lock.

                **Two call patterns, both must be accepted:**

                1. Normal unlock — the CLI sends the full `LockInfo` in the body.
                   The server should compare `ID` to the held lock and refuse on
                   mismatch.
                2. `terraform force-unlock <LOCK_ID>` — the CLI sends an **empty
                   body**; the lock ID is not forwarded. Your server must treat an
                   empty-body unlock as "force-release whatever lock is currently
                   held on this state." This is a long-standing Terraform quirk
                   (see issue #28421).

                Returning 423 here will break `force-unlock`.
            requestBody:
                required: false
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LockInfo'
            responses:
                '200':
                    description: Lock released.
                '204':
                    description: Lock released (no body).
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '409':
                    description: Lock ID in body does not match the held lock.

components:
    parameters:
        StateName:
            name: name
            in: path
            required: true
            description: Workspace / state name. The HTTP backend itself has no
                notion of workspaces, so servers typically put the workspace in the
                path and the client configures one `address` per workspace.
            schema:
                type: string

        LockID:
            name: ID
            in: query
            required: false
            description: Lock ID obtained from the preceding LOCK request. Present
                on state writes whenever locking is enabled on the client.
            schema:
                type: string
                format: uuid

    responses:
        Unauthorized:
            description: Missing or invalid credentials.
        Forbidden:
            description: Authenticated but not permitted to access this state.
        BadRequest:
            description: Malformed request (invalid JSON, missing lock ID, etc.).

    schemas:
        TerraformState:
            type: object
            description: |
                Raw Terraform state JSON. The server should treat this as an
                opaque blob — do not validate or rewrite it. The shape evolves
                across Terraform versions; relevant top-level fields as of
                state format v4 are listed below for reference only.
            additionalProperties: true
            properties:
                version:
                    type: integer
                    description: State file format version (4 as of Terraform 0.12+).
                terraform_version:
                    type: string
                    description: Version of Terraform that wrote this state.
                serial:
                    type: integer
                    description: Monotonic counter incremented on each write.
                lineage:
                    type: string
                    format: uuid
                    description: Random UUID generated on first write; stable for the life of the state.
                outputs:
                    type: object
                    additionalProperties: true
                resources:
                    type: array
                    items:
                        type: object
                        additionalProperties: true
                check_results:
                    type: array
                    nullable: true
                    items:
                        type: object
                        additionalProperties: true

        LockInfo:
            type: object
            description: |
                Matches the Go `statemgr.LockInfo` struct. Field names are
                **PascalCase** because Go's default JSON marshaling is used —
                do not lowercase them.
            required: [ID, Operation, Who, Version, Created]
            properties:
                ID:
                    type: string
                    format: uuid
                    description: Client-generated UUID identifying this lock attempt.
                    example: 2f3b1c4e-6e8f-4b8c-9d2f-1a2b3c4d5e6f
                Operation:
                    type: string
                    description: |
                        Terraform operation requesting the lock. Informational only.
                        Known values include `OperationTypeInvalid`, `OperationTypePlan`,
                        `OperationTypeApply`, and `OperationTypeRefresh`.
                    example: OperationTypeApply
                Info:
                    type: string
                    description: Free-form additional info. Usually empty.
                    default: ''
                Who:
                    type: string
                    description: '`user@hostname` of the client acquiring the lock.'
                    example: alice@workstation
                Version:
                    type: string
                    description: Terraform version string of the client.
                    example: '1.9.2'
                Created:
                    type: string
                    format: date-time
                    description: UTC timestamp in RFC3339Nano format (Go's default time marshaling).
                    example: '2026-04-20T10:00:00.123456789Z'
                Path:
                    type: string
                    description: >
                        Used by the local backend to record the state file path. Usually
                        empty for remote backends; store and return verbatim.
                    default: ''